+++ Changelog.md
@@ -1,10 +1,15 @@
 # Change log
 
 ### vNext
-- Feature: Support apollo-client 2.0
-- Fix: Scope query recyclers by client [#876](https://github.com/apollographql/react-apollo/pull/876)
+- Fix: ensure queryRecycler exists before using it
 - MockNetworkInterface match mock requests regardless of variable order [#973](https://github.com/apollographql/react-apollo/pull/973)
 - Allow to pass removeTypenames to MockedProvider [#1001](https://github.com/apollographql/react-apollo/pull/1001)
+
+### 1.4.14
+- Fix: Scope query recyclers by client [#876](https://github.com/apollographql/react-apollo/pull/876)
+
+### 1.4.13 [DEPRECATED]
+- Support apollo-client 2.0
 
 ### 1.4.12
 - Fix: fix issue with bad deploy